backend/requirements.txt
Flask==3.0.*
flask-cors==4.0.*
requests==2.*
python-dotenv==1.*
